[Feat/#102]: 로그인 상태 초기값 설정 및 스토어명 변경
