+++ src/components/Header/Navigation.tsx
@@ -1,11 +1,11 @@
 import PATH from '@/constants/path';
-import useSessionStore from '@/store/sessionStore';
+import useLoginStore from '@/store/loginStore';
 import { CartIcon } from 'public/assets/icons';
 import { Link } from 'react-router-dom';
 import styled from 'styled-components';
 
 function Navigation() {
-  const isLoggedIn = useSessionStore((state) => state.isLoggedIn);
+  const isLoggedIn = useLoginStore((state) => state.isLoggedIn);
 
   return (
     <div>
