+++ src/hooks/logout/useLogOut.ts
@@ -1,17 +1,16 @@
-import { STORAGE_KEY } from '@/constants/storage';
-import useSessionStore from '@/store/sessionStore';
+import PATH from '@/constants/path';
+import useLoginStore from '@/store/loginStore';
+import { deleteAllFromStorage } from '@/utils/storage';
 import { useNavigate } from 'react-router-dom';
 
 const useLogOut = () => {
   const navigate = useNavigate();
-  const setIsLoggedIn = useSessionStore((state) => state.setIsLoggedIn);
+  const { setIsLoggedIn } = useLoginStore();
 
-  // TODO useType도 삭제
   const handleLogOut = () => {
-    sessionStorage.removeItem(STORAGE_KEY.accessToken);
-    sessionStorage.removeItem(STORAGE_KEY.refreshToken);
+    deleteAllFromStorage();
     setIsLoggedIn(false);
-    navigate('/');
+    navigate(PATH.main);
   };
 
   return handleLogOut;
