+++ src/routes/PublicRoutes.tsx
@@ -1,10 +1,10 @@
 import PATH from '@/constants/path';
 import AuthLayout from '@/layouts/AuthLayout';
-import useSessionStore from '@/store/sessionStore';
+import useLoginStore from '@/store/loginStore';
 import { Navigate } from 'react-router-dom';
 
 function PublicRoutes() {
-  const isLoggedIn = useSessionStore((state) => state.isLoggedIn);
+  const isLoggedIn = useLoginStore((state) => state.isLoggedIn);
   return isLoggedIn ? <Navigate to={PATH.main} /> : <AuthLayout />;
 }
